Do not collect if missing learners characteristics
In previous versions, data would be collected even if the learner
has no specified its age or gender. If this data was collected and
rejected by the server, it would still be backed-up and restore
in the next attempt.

The problem is that the same restored data would be sent to server
and rejected over and over. In other words data would never be
collected again.

Signed-off-by: Martin Abente Lahaye <tch@sugarlabs.org>